+++ app/list/page.jsx
@@ -1,4 +1,4 @@
-import CurrencyTable from '../components/CurrencyTable';
+import CurrencyTable from '../components/antiguo';
 
 export default function Page() {
   return (
